refactor(dbtool): extract shared prover-task admin helpers
The forthcoming EE prover-task admin commands accept the same status
filters, hex-key encoding, --confirm guard, and PermanentFailure
reason string as the OL surface; lifting these to a shared module
keeps the two surfaces in lockstep so operator workflows behave
identically against either store.
