+++ bin/strata-dbtool/src/cmd/mod.rs
@@ -6,5 +6,6 @@ pub(crate) mod l1;
 pub(crate) mod ol;
 pub(crate) mod ol_state;
 pub(crate) mod prover_task;
+pub(crate) mod prover_task_common;
 pub(crate) mod syncinfo;
 pub(crate) mod writer;
